Refactor random joke into seperate component
