+++ joke-app/src/Components/HomePage/HomePage.css
@@ -1,0 +1,8 @@
+.HomePage {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+}
